+++ src/builder/index.ts
@@ -1,0 +1,2 @@
+export * from './typeorm-url.class';
+export * from './uml-builder.class';