+++ src/types/direction.enum.ts
@@ -1,0 +1,4 @@
+export enum Direction {
+	TB = 'TB',
+	LR = 'LR',
+}