+++ src/types/flags.interface.ts
@@ -1,0 +1,12 @@
+export interface Flags {
+	connection?: string,
+	direction?: string,
+	download?: string,
+	echo?: boolean,
+	exclude?: string,
+	format?: string,
+	handwritten?: boolean,
+	include?: string,
+	monochrome?: boolean,
+	['with-enum-values']?: boolean,
+}