+++ src/types/format.enum.ts
@@ -1,0 +1,6 @@
+export enum Format {
+	PNG = 'png',
+	SVG = 'svg',
+	TXT = 'txt',
+	PUML = 'puml',
+}